this is a big one and... risky.
Separated out writing multidata from the threading process
and now check if a seed is beatable /before/ writing any
output. This involved moving rom_names from a threaded process
and into their respective __init__'s across various games.

This branch will be archived for posterity in order to move forward
to post-0.5.0. It is not suggested to use this branch. Whether some
of these changes survive the impending new "personal" branch is TBD.

Also lots of SMZ3 changes, again, some of which will make it into
the new/upcoming "personal" branch, and others will not.